// src/lib/db-types.ts
/**
 * This file was auto-generated by a Deno script.
 * Please do not edit this file directly.
 *
 * Generated at: 2025-10-23T04:42:21.134Z
 */

export interface Account {
  /** @db_type int4 */
  id: number;
  /** @db_type text */
  username: string;
  /** @db_type timestamp */
  created_at: Date;
}

export interface Level {
  /** @db_type int4 */
  id: number;
  /** @db_type text */
  name: string;
  /** @db_type int4 */
  publisher_account_id: number;
  /** @db_type text */
  description?: string | null;
  /** @db_type leveldifficulty */
  difficulty: 'N/A' | 'Auto' | 'Easy' | 'Normal' | 'Hard' | 'Harder' | 'Insane' | 'Easy Demon' | 'Medium Demon' | 'Hard Demon' | 'Insane Demon' | 'Extreme Demon';
  /** @db_type levellength */
  length: 'Tiny' | 'Short' | 'Medium' | 'Long' | 'XL' | 'Platformer';
  /** @db_type int4 */
  song_id: number;
  /** @db_type date */
  release_date?: Date | null;
  /** @db_type text */
  video_url?: string | null;
  /** @db_type timestamp */
  created_at: Date;
  /** @db_type int2 */
  coins: number;
  /** @db_type bool */
  two_player: boolean;
  /** @db_type levelrating */
  rating: 'Rated' | 'Featured' | 'Epic' | 'Legendary' | 'Mythic';
  /** @db_type _int4 */
  creators_account_ids?: number[] | null;
  /** @db_type _int4 */
  skillsets?: number[] | null;
}

export interface Progress {
  /** @db_type int4 */
  id: number;
  /** @db_type int4 */
  user_id: number;
  /** @db_type int4 */
  level_id: number;
  /** @db_type progressstatus */
  status: 'To Try' | 'In Progress' | 'Completed' | 'Dropped';
  /** @db_type int2 */
  completion_pct?: number | null;
  /** @db_type int4 */
  completion_time?: number | null;
  /** @db_type int4 */
  total_attempts?: number | null;
  /** @db_type date */
  start_date?: Date | null;
  /** @db_type date */
  completion_date?: Date | null;
  /** @db_type int2 */
  enjoyment_rating?: number | null;
  /** @db_type text */
  video_url?: string | null;
  /** @db_type text */
  review?: string | null;
  /** @db_type timestamp */
  created_at: Date;
  /** @db_type timestamp */
  updated_at: Date;
  /** @db_type int4 */
  placement?: number | null;
}

export interface Skillset {
  /** @db_type int4 */
  id: number;
  /** @db_type text */
  name: string;
}

export interface Song {
  /** @db_type int4 */
  id: number;
  /** @db_type text */
  title: string;
  /** @db_type text */
  artist: string;
  /** @db_type timestamp */
  created_at: Date;
}

export interface Tag {
  /** @db_type int4 */
  id: number;
  /** @db_type text */
  name: string;
}

export interface User {
  /** @db_type int4 */
  id: number;
  /** @db_type text */
  username: string;
  /** @db_type text */
  email: string;
  /** @db_type text */
  password_hash: string;
  /** @db_type text */
  bio?: string | null;
  /** @db_type text */
  profile_picture_url?: string | null;
  /** @db_type timestamp */
  created_at: Date;
  /** @db_type _userrole */
  roles: ('Owner' | 'Admin' | 'User')[];
}